first fix from my new zend_parse_parameters_* checker

+++ ext/mbstring/mbstring.c
@@ -17,7 +17,7 @@
    +----------------------------------------------------------------------+
  */
 
-/* $Id: mbstring.c,v 1.224.2.22.2.12 2006/08/31 16:15:24 tony2001 Exp $ */
+/* $Id: mbstring.c,v 1.224.2.22.2.13 2006/09/05 12:30:35 nlopess Exp $ */
 
 /*
  * PHP 4 Multibyte String module "mbstring"
@@ -2824,7 +2824,7 @@ PHP_FUNCTION(mb_encode_mimeheader)
 	int trans_enc_name_len;
 	char *linefeed = "\r\n";
 	int linefeed_len;
-	int indent = 0;
+	long indent = 0;
 
 	mbfl_string_init(&string);
 	string.no_language = MBSTRG(current_language);
